pyproject.toml
[tool.poetry]
name = "pl_train"
version = "0.1.0"
description = ""
authors = ["sarrrrry <sarrrrry.sarrrrry@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytorch-lightning = "^1.5.10"
torchmetrics = "^0.7.2"
matplotlib = "^3.5.1"
torchvision = "^0.11.3"
wandb = "^0.12.10"
hydra-core = "^1.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
